Doc tool update
Automatic detection of the inverter type. Update needed for the 03 packets from MIX / SPA / SPH as they use common register set here. SPH and MIX use the same registers in 04 and at the moment this packets will be detected as SPH.
Updated examples in the file
Navigation in packet should match 1:1 to the description in the protocol documentation. The registers as described in the documentation are translated internally to the corresponding positions in the data packet.

Signed-off-by: rany <ranygh@riseup.net>